requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
colorama==0.4.6
cramjam==2.10.0
Django==5.2
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-phonenumber-field==8.1.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docx==0.2.4
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
lxml==5.4.0
mongoengine==0.29.1
numpy==2.2.5
packaging==24.2
phonenumbers==9.0.3
pillow==11.2.1
pluggy==1.5.0
PyJWT==2.9.0
pymongo==3.11.4
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
python-docx==1.1.2
python-snappy==0.7.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
winkerberos==0.12.2
